Apply dynamic imports to TimesSquarePrGitHubNav
- Created TimesSquarePrGitHubNavClient.tsx for client-only rendering
- Updated TimesSquarePrGitHubNav.tsx to use dynamic import with ssr:
  false
- Prevents useGitHubPrContentsListing SWR hook from executing during SSR
- Maintains full client-side PR navigation functionality including PR
  badges and checks

This resolves potential SSR issues with the GitHub PR contents listing
hook that fetches pull request data, check status, and repository
navigation from the Times Square API for GitHub PR preview pages.
