+++ apps/squareone/src/components/TimesSquarePrGitHubNav/TimesSquarePrGitHubNav.tsx
@@ -1,12 +1,9 @@
-import React from 'react';
-import styled from 'styled-components';
-import getConfig from 'next/config';
+/*
+ * TimesSquarePrGitHubNav with dynamic import to prevent SSR issues
+ * Uses client-only component to handle SWR hooks safely.
+ */
 
-import TimesSquareGitHubNav from '../TimesSquareGitHubNav';
-import useGitHubPrContentsListing from './useGitHubPrContentsListing';
-import GitHubPrTitle from './GitHubPrTitle';
-import GitHubPrBadge from './GitHubPrBadge';
-import GitHubCheckBadge from './GitHubCheckBadge';
+import dynamic from 'next/dynamic';
 
 type TimesSquarePrGitHubNavProps = {
   owner: string;
@@ -15,92 +12,29 @@ type TimesSquarePrGitHubNavProps = {
   showPrDetails?: boolean;
 };
 
+// Dynamic import with SSR disabled to prevent SWR hook issues
+const TimesSquarePrGitHubNavClient = dynamic(
+  () => import('./TimesSquarePrGitHubNavClient'),
+  {
+    ssr: false,
+    loading: () => null, // No loading state needed for navigation
+  }
+);
+
 function TimesSquarePrGitHubNav({
   owner,
   repo,
   commitSha,
   showPrDetails = true,
 }: TimesSquarePrGitHubNavProps) {
-  const { publicRuntimeConfig } = getConfig();
-  const { timesSquareUrl } = publicRuntimeConfig;
-  const githubContents = useGitHubPrContentsListing(
-    timesSquareUrl,
-    owner,
-    repo,
-    commitSha
+  return (
+    <TimesSquarePrGitHubNavClient
+      owner={owner}
+      repo={repo}
+      commitSha={commitSha}
+      showPrDetails={showPrDetails}
+    />
   );
-
-  if (!(githubContents.loading || githubContents.error)) {
-    const { nbCheck, yamlCheck } = githubContents;
-
-    return (
-      <StyledSection>
-        {showPrDetails && (
-          <>
-            <GitHubPrTitle owner={owner} repo={repo} commit={commitSha} />
-
-            <ItemList>
-              {githubContents.pullRequests.map((pr) => (
-                <li key={`pr-${pr.number}`}>
-                  <GitHubPrBadge
-                    state={pr.state}
-                    number={pr.number}
-                    url={pr.conversation_url}
-                    title={pr.title}
-                    authorName={pr.contributor.username}
-                    authorAvatarUrl={pr.contributor.avatar_url}
-                    authorUrl={pr.contributor.html_url}
-                  />
-                </li>
-              ))}
-            </ItemList>
-
-            <ItemList>
-              {nbCheck && (
-                <li>
-                  <GitHubCheckBadge
-                    status={nbCheck.status}
-                    title={nbCheck.report_title}
-                    conclusion={nbCheck.conclusion}
-                    url={nbCheck.html_url}
-                  />
-                </li>
-              )}
-              {yamlCheck && (
-                <li>
-                  <GitHubCheckBadge
-                    status={yamlCheck.status}
-                    title={yamlCheck.report_title}
-                    conclusion={yamlCheck.conclusion}
-                    url={yamlCheck.html_url}
-                  />
-                </li>
-              )}
-            </ItemList>
-          </>
-        )}
-
-        <h3>Notebooks</h3>
-        <TimesSquareGitHubNav
-          contentNodes={githubContents.contents}
-          pagePathRoot="/times-square/github-pr"
-          pagePath={null}
-        />
-      </StyledSection>
-    );
-  }
-
-  return null; // still loading
 }
 
 export default TimesSquarePrGitHubNav;
-
-const ItemList = styled.ul`
-  list-style: none;
-  margin: 1rem 0 1rem;
-  padding-left: 0;
-`;
-
-const StyledSection = styled.section`
-  margin-top: 2rem;
-`;
